impelement parallax
